test: consolidate prefer-destructuring tests
